store/tikv: share type OneByOneSuite for tests related to tikv (#25098)
