+++ store/driver/util_test.go
@@ -17,7 +17,6 @@ import (
 	"context"
 	"flag"
 	"fmt"
-	"sync"
 	"testing"
 
 	. "github.com/pingcap/check"
@@ -26,6 +25,7 @@ import (
 	"github.com/pingcap/tidb/store/copr"
 	"github.com/pingcap/tidb/store/mockstore/unistore"
 	"github.com/pingcap/tidb/store/tikv"
+	"github.com/pingcap/tidb/store/tikv/mockstore"
 )
 
 func TestT(t *testing.T) {
@@ -34,9 +34,7 @@ func TestT(t *testing.T) {
 }
 
 var (
-	withTiKVGlobalLock sync.RWMutex
-	WithTiKV           = flag.Bool("with-tikv", false, "run tests with TiKV cluster started. (not use the mock server)")
-	pdAddrs            = flag.String("pd-addrs", "127.0.0.1:2379", "pd addrs")
+	pdAddrs = flag.String("pd-addrs", "127.0.0.1:2379", "pd addrs")
 )
 
 // NewTestStore creates a kv.Storage for testing purpose.
@@ -45,7 +43,7 @@ func NewTestStore(c *C) kv.Storage {
 		flag.Parse()
 	}
 
-	if *WithTiKV {
+	if *mockstore.WithTiKV {
 		var d TiKVDriver
 		store, err := d.Open(fmt.Sprintf("tikv://%s", *pdAddrs))
 		c.Assert(err, IsNil)
@@ -82,20 +80,4 @@ func clearStorage(store kv.Storage) error {
 }
 
 // OneByOneSuite is a suite, When with-tikv flag is true, there is only one storage, so the test suite have to run one by one.
-type OneByOneSuite struct{}
-
-func (s *OneByOneSuite) SetUpSuite(c *C) {
-	if *WithTiKV {
-		withTiKVGlobalLock.Lock()
-	} else {
-		withTiKVGlobalLock.RLock()
-	}
-}
-
-func (s *OneByOneSuite) TearDownSuite(c *C) {
-	if *WithTiKV {
-		withTiKVGlobalLock.Unlock()
-	} else {
-		withTiKVGlobalLock.RUnlock()
-	}
-}
+type OneByOneSuite = mockstore.OneByOneSuite
